# swagger/openapi.yaml
openapi: 3.0.0
info:
  title: The race cars
  contact:
    email: loi.tra@9thwonder.com
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/
  - url: https://virtserver.swaggerhub.com/annguyen9w/hapi-starter/1.0.0
    description: SwaggerHub API Auto Mocking
paths:
  /classes:
    get:
      tags:
        - Class
      summary: Get all class
      operationId: getClasses
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Class'
        # '400':
        #   $ref: '#/components/responses/badrequest'
    post:
      tags:
        - Class
      summary: Add a new class to the system
      operationId: addClass
      requestBody:
        $ref: '#/components/requestBodies/Class'
      responses:
        '200':
          $ref: '#/components/responses/successwithid'
        '400':
          $ref: '#/components/responses/badrequest'
  /classes/{classId}:
    get:
      tags:
        - Class
      summary: Find class by ID
      operationId: getClassById
      parameters:
        - name: classId
          in: path
          description: ID of class to return
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Class'
        '400':
          $ref: '#/components/responses/badrequest'
        '404':
          $ref: '#/components/responses/notfound'
    put:
      tags:
        - Class
      summary: Update an existing class
      operationId: updateClass
      parameters:
        - name: classId
          in: path
          description: Class id need to be updated
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        $ref: '#/components/requestBodies/Class'
      responses:
        '200':
          $ref: '#/components/responses/success'
        '400':
          $ref: '#/components/responses/badrequest'
        '404':
          $ref: '#/components/responses/notfound'
    delete:
      tags:
        - Class
      summary: Delete a class
      operationId: deleteClass
      parameters:
        - name: classId
          in: path
          description: Class id to delete
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          $ref: '#/components/responses/success'
        '400':
          $ref: '#/components/responses/badrequest'
        '404':
          $ref: '#/components/responses/notfound'
  /addresses:
    get:
      tags:
        - Address
      summary: Get all address
      operationId: getAddresses
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Address'
        # '400':
        #   $ref: '#/components/responses/badrequest'
    post:
      tags:
        - Address
      summary: Add a new address to the system
      operationId: addAddress
      requestBody:
        $ref: '#/components/requestBodies/Address'
      responses:
        '200':
          $ref: '#/components/responses/successwithid'
        '400':
          $ref: '#/components/responses/badrequest'
  /addresses/{addressId}:
    get:
      tags:
        - Address
      summary: Find address by ID
      operationId: getAddressById
      parameters:
        - name: addressId
          in: path
          description: ID of address to return
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
        '400':
          $ref: '#/components/responses/badrequest'
        '404':
          $ref: '#/components/responses/notfound'
    put:
      tags:
        - Address
      summary: Update an existing address
      operationId: updateAddress
      parameters:
        - name: addressId
          in: path
          description: Address id need to be updated
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        $ref: '#/components/requestBodies/Address'
      responses:
        '200':
          $ref: '#/components/responses/success'
        '400':
          $ref: '#/components/responses/badrequest'
        '404':
          $ref: '#/components/responses/notfound'
    delete:
      tags:
        - Address
      summary: Delete a address
      operationId: deleteAddress
      parameters:
        - name: addressId
          in: path
          description: Address id to delete
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          $ref: '#/components/responses/success'
        '400':
          $ref: '#/components/responses/badrequest'
        '404':
          $ref: '#/components/responses/notfound'
...
  /cars:
    get:
      tags:
        - Car
      summary: Get all car
      operationId: getCars
      parameters:
        - name: make
          in: query
          description: Make values that need to be considered for filter
          required: false
          schema:
            type: string
            example: Ferrari
        - name: model
          in: query
          description: Model values that need to be considered for filter
          required: false
          schema:
            type: string
            example: 488 GTE-LM
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CarResponse'
        '400':
          $ref: '#/components/responses/badrequest'
    post:
      tags:
        - Car
      summary: Add a new car to the system
      operationId: addCar
      requestBody:
        $ref: '#/components/requestBodies/Car'
      responses:
        '200':
          $ref: '#/components/responses/successwithid'
        '400':
          $ref: '#/components/responses/badrequest'
  /cars/{carId}:
    get:
      tags:
        - Car
      summary: Find car by ID
      operationId: getCarById
      parameters:
        - name: carId
          in: path
          description: ID of car to return
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarResponse'
        '400':
          $ref: '#/components/responses/badrequest'
        '404':
          $ref: '#/components/responses/notfound'
    put:
      tags:
        - Car
      summary: Update an existing car
      operationId: updateCar
      parameters:
        - name: carId
          in: path
          description: Car id need to be updated
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        $ref: '#/components/requestBodies/Car'
      responses:
        '200':
          $ref: '#/components/responses/success'
        '400':
          $ref: '#/components/responses/badrequest'
        '404':
          $ref: '#/components/responses/notfound'
    delete:
      tags:
        - Car
      summary: Delete a car
      operationId: deleteCar
      parameters:
        - name: carId
          in: path
          description: Car id to delete
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          $ref: '#/components/responses/success'
        '400':
          $ref: '#/components/responses/badrequest'
        '404':
          $ref: '#/components/responses/notfound'
  /cars/{carId}/results:
    get:
      tags:
        - Car
      summary: All race results for that car
      operationId: getRaceResultsByCarId
      parameters:
        - name: carId
          in: path
          description: ID of car to return
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RaceResult'
        '400':
          $ref: '#/components/responses/badrequest'
        '404':
          $ref: '#/components/responses/notfound'
...
components:
  schemas:
    UUID:
      type: string
      format: uuid
      example: 40e6215d-b5c6-4896-987c-f30f3678f608
    ClassRequestBody:
      required:
        - name
      type: object
      properties:
        name:
          type: string
          example: LM GTE AM
    Class:
      allOf:
        - required:
            - id
          type: object
          properties:
            id:
              $ref: '#/components/schemas/UUID'
        - $ref: '#/components/schemas/ClassRequestBody'
    AddressRequestBody:
      required:
        - city
        - country
        - state
        - zipcode
      type: object
      properties:
        street:
          type: string
          example: 1797 Colonial Drive
        street2:
          type: string
          example: 411 Patterson Street
        city:
          type: string
          example: Houston
        state:
          type: string
          example: Texas
        zipcode:
          type: string
          example: '77002'
        country:
          type: string
          example: United States
    Address:
      allOf:
        - required:
            - id
          type: object
          properties:
            id:
              $ref: '#/components/schemas/UUID'
        - $ref: '#/components/schemas/AddressRequestBody'
...
    Driver:
      required:
        - firstName
        - id
        - lastName
        - nationality
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        firstName:
          type: string
          example: Nguyen
        lastName:
          type: string
          example: An
        nationality:
          type: string
          enum:
            - USA
            - Viet Nam
        homeAddress:
          $ref: '#/components/schemas/Address'
        managementAddress:
          $ref: '#/components/schemas/Address'
...
    Team:
      required:
        - cars
        - id
        - name
        - nationality
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          example: 7 TOYOTA GAZOO RACING
        nationality:
          type: string
          enum:
            - USA
            - Viet Nam
        businessAddress:
          $ref: '#/components/schemas/Address'
...
    CarRequestBody:
      required:
        - class
        - make
        - model
        - team
      type: object
      properties:
        make:
          type: string
          example: Ferrari
        model:
          type: string
          example: 488 GTE-LM
        class:
          $ref: '#/components/schemas/UUID'
        team:
          $ref: '#/components/schemas/UUID'
    Car:
      required:
        - class
        - make
        - model
        - team
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        make:
          type: string
          example: Ferrari
        model:
          type: string
          example: 488 GTE-LM
        class:
          $ref: '#/components/schemas/Class'
    CarResponse:
      allOf:
        - $ref: '#/components/schemas/Car'
        - required:
            - team
          type: object
          properties:
            team:
              $ref: '#/components/schemas/Team'
...
    Race:
      required:
        - name
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          example: Le Mans
...
    RaceResult:
      required:
        - car
        - class
        - driver
        - id
        - race
        - raceNumber
        - startPosition
      type: object
      properties:
        id:
          type: string
          format: uuid
        race:
          $ref: '#/components/schemas/Race'
        car:
          $ref: '#/components/schemas/Car'
        driver:
          $ref: '#/components/schemas/Driver'
        class:
          $ref: '#/components/schemas/Class'
        raceNumber:
          type: string
          example: '200'
        startPosition:
          minimum: 1
          type: integer
          example: 2
        finishPosition:
          minimum: 1
          type: integer
          example: 3
  requestBodies:
    Class:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ClassRequestBody'
      required: true
    Address:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AddressRequestBody'
      required: true
...
    Car:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CarRequestBody'
      required: true
...
  responses:
    success:
      description: Successful operation
      content:
        text/plain:
          schema:
            type: string
            example: success
    successwithid:
      description: Successful operation
      content:
        application/json:
          schema:
            type: object
            properties:
              id: 
                $ref: '#/components/schemas/UUID'
              status:
                type: string
                default: success
            required:
              - id
              - status
    badrequest:
      description: Invalid input
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode: 
                type: integer
                default: 400
              error:
                type: string
                default: Bad Request
              message:
                type: string
                default: Invalid request input
            required:
              - statusCode
              - error
              - message
    notfound:
      description: Item not found
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode: 
                type: integer
                default: 404
              error:
                type: string
                default: Not Found
              message:
                type: string
                default: Not Found
            required:
              - statusCode
              - error
              - message